requirements.txt
cffi==1.15.0
matplotlib==3.1.2
numpy==1.17.4
pandas==0.25.3
PyYAML==6.0
scipy==1.3.3
